Move all states to private module
