+++ src/poliastro/twobody/orbit.py
@@ -10,7 +10,6 @@ from astropy.coordinates import (
     CartesianRepresentation,
     get_body_barycentric_posvel,
 )
-
 from astroquery.jplhorizons import Horizons
 
 from poliastro.bodies import Earth, Moon, Sun
@@ -18,11 +17,10 @@ from poliastro.constants import J2000
 from poliastro.core.angles import nu_to_M as nu_to_M_fast
 from poliastro.core.elements import rv2coe
 from poliastro.frames import Planes, get_frame
-from poliastro.twobody import classical, equinoctial, rv
 from poliastro.twobody.angles import E_to_nu, nu_to_M
 from poliastro.twobody.propagation import mean_motion, propagate
 
-from ._base import BaseState  # flake8: noqa
+from ._states import BaseState, ClassicalState, ModifiedEquinoctialState, RVState
 
 ORBIT_FORMAT = "{r_p:.0f} x {r_a:.0f} x {inc:.1f} ({frame}) orbit around {body} at epoch {epoch} ({scale})"
 # String representation for orbits around bodies without predefined reference frame
@@ -110,7 +108,7 @@ class Orbit(object):
         """
         assert np.any(r.value), "Position vector must be non zero"
 
-        ss = rv.RVState(attractor, r, v)
+        ss = RVState(attractor, r, v)
         return cls(ss, epoch, plane)
 
     @classmethod
@@ -160,9 +158,7 @@ class Orbit(object):
         if ecc > 1 and a > 0:
             raise ValueError("Hyperbolic orbits have negative semimajor axis")
 
-        ss = classical.ClassicalState(
-            attractor, a * (1 - ecc ** 2), ecc, inc, raan, argp, nu
-        )
+        ss = ClassicalState(attractor, a * (1 - ecc ** 2), ecc, inc, raan, argp, nu)
         return cls(ss, epoch, plane)
 
     @classmethod
@@ -194,7 +190,7 @@ class Orbit(object):
             Fundamental plane of the frame.
 
         """
-        ss = equinoctial.ModifiedEquinoctialState(attractor, p, f, g, h, k, L)
+        ss = ModifiedEquinoctialState(attractor, p, f, g, h, k, L)
         return cls(ss, epoch, plane)
 
     @classmethod
@@ -324,7 +320,7 @@ class Orbit(object):
             Fundamental plane of the frame.
 
         """
-        ss = classical.ClassicalState(attractor, p, 1.0 * u.one, inc, raan, argp, nu)
+        ss = ClassicalState(attractor, p, 1.0 * u.one, inc, raan, argp, nu)
         return cls(ss, epoch, plane)
 
     def represent_as(self, representation):
